new conversion table

+++ src/serial.h
@@ -5,6 +5,6 @@
 
 bool serial_read (int fd);
 int serial_open (void);
-void serial_write (int fd, int a, int b, int c);
+void serial_write (int fd, double a, double b, double c);
 
 #endif /* _serial_h_ */
